CHANGE: big code restructure in image analysis example
